+++ src/talkpipe/llm/prompt_adapters.py
@@ -1,5 +1,6 @@
 from .prompt_adapter_base import AbstractLLMPromptAdapter, logger
 from .prompt_adapters_anthropic import AnthropicPromptAdapter
+from .prompt_adapters_eliza import ElizaPromptAdapter
 from .prompt_adapters_ollama import OllamaPromptAdapter
 from .prompt_adapters_openai import OpenAIPromptAdapter
 
@@ -9,5 +10,6 @@ __all__ = [
     "OllamaPromptAdapter",
     "AnthropicPromptAdapter",
     "OpenAIPromptAdapter",
+    "ElizaPromptAdapter",
     "logger",
 ]